get hash endpoint & middleware organization
