+++ routes/token.js
@@ -1,3 +1,3 @@
-const token = 'eyJhbGciOiJSUzI1NiIsImtpZCI6IjQ5MjcxMGE3ZmNkYjE1Mzk2MGNlMDFmNzYwNTIwYTMyYzg0NTVkZmYiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJhY2NvdW50cy5nb29nbGUuY29tIiwiYXpwIjoiOTk2OTk3OTMyNDUzLTJnaGE1Njg1cjRpZW0yMjdjYmM0dmpmcDFzbWUxazJ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiYXVkIjoiOTk2OTk3OTMyNDUzLTJnaGE1Njg1cjRpZW0yMjdjYmM0dmpmcDFzbWUxazJ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwic3ViIjoiMTA2NTAxMTYyNTc4OTk2NDQzNzE2IiwiZW1haWwiOiJ2YW5zaGlrYXB1bmRpcjkyQGdtYWlsLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJhdF9oYXNoIjoialV3bmx2RHQ1S3lVS3Nmb0h1V2dDQSIsIm5hbWUiOiJWYW5zaGlrYSBQdW5kaXIiLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDUuZ29vZ2xldXNlcmNvbnRlbnQuY29tLy1JY1g0cWdVMGxYTS9BQUFBQUFBQUFBSS9BQUFBQUFBQUFBQS9BTVp1dWNteGFqWmFsUGUxeDFFdjN6Q0tfa2dvaHJ3SGh3L3M5Ni1jL3Bob3RvLmpwZyIsImdpdmVuX25hbWUiOiJWYW5zaGlrYSIsImZhbWlseV9uYW1lIjoiUHVuZGlyIiwibG9jYWxlIjoiZW4iLCJpYXQiOjE1OTE1NzM1MTAsImV4cCI6MTU5MTU3NzExMCwianRpIjoiYzcwMjQ0NDI3OWE2N2JmYTkyNzgzZDg5ZmQ0NGIzY2FjMzE3ZmNmNyJ9.j-CtXaxzl5ZcSKEVWPoy2DmEPwe3OKZ2MaX3zKodqb0hzUXB7Jg1gex9lMrF70b8HUmnnCYXAk8lLhJ8u4HbaRbeVHwBlgZadVxpEa_9BaCpLwAt7Nm71v3nf66GtdiH0Zil2XkWdxi5okAc9uawM3-8JUnlHzATT1C0QNaIJLdgT_KPZS7xAeBbWYyfiw940C4WNWQrr_Lb8KJ8FzddV6Re59wjfs0Iqq0pMlHXYf1-ZueoqupVevrI50iQDMOJ4vRJLJlHLnfxgBd-ww-hQa1x5NsyntCRWk_z9KYXA2KQZZFTtZVTlJUIJwqdwoz9XO411s0_0PgjAH7wlhDkyg';
+const token = 'eyJhbGciOiJSUzI1NiIsImtpZCI6IjQ5MjcxMGE3ZmNkYjE1Mzk2MGNlMDFmNzYwNTIwYTMyYzg0NTVkZmYiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJhY2NvdW50cy5nb29nbGUuY29tIiwiYXpwIjoiOTk2OTk3OTMyNDUzLTJnaGE1Njg1cjRpZW0yMjdjYmM0dmpmcDFzbWUxazJ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiYXVkIjoiOTk2OTk3OTMyNDUzLTJnaGE1Njg1cjRpZW0yMjdjYmM0dmpmcDFzbWUxazJ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwic3ViIjoiMTA2NTAxMTYyNTc4OTk2NDQzNzE2IiwiZW1haWwiOiJ2YW5zaGlrYXB1bmRpcjkyQGdtYWlsLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJhdF9oYXNoIjoiT1k2UURpLWhJMmxLZDVEZUVNa1J5ZyIsIm5hbWUiOiJWYW5zaGlrYSBQdW5kaXIiLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDUuZ29vZ2xldXNlcmNvbnRlbnQuY29tLy1JY1g0cWdVMGxYTS9BQUFBQUFBQUFBSS9BQUFBQUFBQUFBQS9BTVp1dWNteGFqWmFsUGUxeDFFdjN6Q0tfa2dvaHJ3SGh3L3M5Ni1jL3Bob3RvLmpwZyIsImdpdmVuX25hbWUiOiJWYW5zaGlrYSIsImZhbWlseV9uYW1lIjoiUHVuZGlyIiwibG9jYWxlIjoiZW4iLCJpYXQiOjE1OTE2MjczNDAsImV4cCI6MTU5MTYzMDk0MCwianRpIjoiNmM0OWI0ZmYxMmZiNDlmMTkzYzI0MjIzMGJkZWU5NmJjOGRhZTRhNiJ9.RD1_3Z6-UlypfyId4SNGY85mg7nX3eprZ27gAI4jHgVen-6LOB41x0iKdZWjHyGjokZSz6mPExupWvOkwu36s0LMMu3_7BddiNe0XSHU6xLqvqrfsYjdTaGOAmjU49udsyUDaPk69p22WGhbtmXomlZURk8hkMqgF2OXOrqLBAyeg9BJkK5b9yK45zAYYxYi7Xcrxso6cli6LfzTBiAHJmAlSOcfdxOaHNrl4OXqbeEIqeCAZtLirzt91Zu6DYgJqKKuNf9utpKwKKFVShprPweM2NrcGisra7pQpsQESQjuwQiJlQT3QW8muj124qdIYrRZrfUmOkol_WVnG_oszQ';
 
 module.exports = token;
